requirements.txt
sentence-transformers==2.2.2
flask==3.0.3
pypdf==4.2.0
python-dotenv==1.0.1

langchain==0.2.16
langchain-core==0.2.43
langchain-community==0.2.16
langchain-pinecone==0.1.3
langchain-experimental==0.0.59
langchain-groq==0.1.4

pinecone==5.0.0
openai==1.40.0
-e .
